Move fuzz to it's own test file to pass build on lower go versions. Add a regex length limit to increase test speed and ignore timeout on very long input.
